backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.109.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.82.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
slowapi==0.1.9
sniffio==1.3.1
starlette==0.36.3
tqdm==4.67.1
typing_extensions==4.13.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
